LPD-90288 Move the FIPS algorithm validation into FIPSModeValidator
The algorithm and key checks lived in a separate FIPSModeUtil while
FIPSModeValidator already owned the FIPS provider validation. Folding
the two together leaves a single entry point for every FIPS validation
and drops the redundant utility class.
